npc: Configurable max learn level
Restrict how high an NPC can advance in any skill during the learning
phase.  This prevents the NPC from using all of its skill points on only
a few skills.  Instead, allow these skills to improve during the mastery
phase.

This setting is specified with `--npc-learn-level-max`.  The default
value is 2.